+++ include/libcotask/task_macros.h
@@ -17,6 +17,11 @@ LIBCOPP_COTASK_NAMESPACE_BEGIN
 struct LIBCOPP_COTASK_API_HEAD_ONLY macro_coroutine {
   using stack_allocator_type = LIBCOPP_COPP_NAMESPACE_ID::allocator::default_statck_allocator;
   using coroutine_type = LIBCOPP_COPP_NAMESPACE_ID::coroutine_context_container<stack_allocator_type>;
-  using data_type = int;
+  using type = int;
+};
+
+template <class T>
+struct LIBCOPP_COPP_API_HEAD_ONLY task_data_ptr_selector {
+  using type = typename LIBCOPP_COPP_NAMESPACE_ID::future::poll_storage_ptr_selector<T>::type;
 };
 LIBCOPP_COTASK_NAMESPACE_END