+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 1. Add ABI tag for all c++ codes
   > We will keep ABI compatitablity for all assemblly APIs and will change the suffix when ABI changed.
-2. \[TODO\] Change the return type to `cotask::task_action_future<T>`
+2. \[TODO\] Change the return type to `cotask::impl::task_action_future<T, TPTR>` (aka `cotask::task<T>::future_type`)
   + Add `data_type` to task_macro ( `macro_coroutine` )
   + Add API `task_action_impl::get_task_id`
   + Add internal API`task_action_impl::get_task`
   + Store result data into `cotask::task<...>`
+3. \[Previous\] Redesign C++ 20 coroutine support.
 
 ## 1.5.0